requirements.txt
requests~=2.31.0
pandas~=2.2.2
halo~=0.0.31
python-telegram-bot~=21.1.1
yfinance~=0.2.38
matplotlib~=3.8.4